+++ packages/plugin-audio/web/src/audio-component.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import AudioViewer from './audio-viewer';
+import type { AudioPluginEditorConfig } from './types';
+import { AUDIO_TYPE } from './types';
+import { DEFAULTS } from './defaults';
+import type { ComponentData, RichContentTheme } from 'wix-rich-content-common';
+
+interface Props {
+  componentData: ComponentData;
+  settings: AudioPluginEditorConfig;
+  theme: RichContentTheme;
+}
+
+class AudioComponent extends React.Component<Props> {
+  static type = { AUDIO_TYPE };
+
+  render() {
+    const { componentData, settings, theme } = this.props;
+    return <AudioViewer componentData={componentData} settings={settings} theme={theme} />;
+  }
+}
+
+export { AudioComponent as Component, DEFAULTS };