+++ packages/plugin-audio/web/src/index.ts
@@ -1,0 +1,4 @@
+export { createAudioPlugin } from './createAudioPlugin';
+export { default as audioViewer } from './audio-viewer';
+export { AUDIO_TYPE } from './types';
+export { pluginAudio } from './editor';