+++ packages/plugin-audio/web/src/toolbar/createToolbar.ts
@@ -1,0 +1,22 @@
+import createInlineButtons from './inline-buttons';
+import createInsertButtons from './insert-buttons';
+import type { CreatePluginToolbar, TranslationFunction } from 'wix-rich-content-common';
+import type { AudioPluginEditorConfig } from '../types';
+
+const createToolbar: CreatePluginToolbar = ({
+  t,
+  settings,
+  isMobile,
+}: {
+  t: TranslationFunction;
+  settings: AudioPluginEditorConfig;
+  isMobile: boolean;
+}) => {
+  return {
+    InlineButtons: createInlineButtons(),
+    InsertButtons: createInsertButtons({ t, settings, isMobile }),
+    name: 'audio',
+  };
+};
+
+export default createToolbar;