+++ packages/plugin-audio/web/src/typeMapper.ts
@@ -1,0 +1,7 @@
+import AudioViewer from './audio-viewer';
+import { AUDIO_TYPE } from './types';
+import type { PluginTypeMapper } from 'wix-rich-content-common';
+
+export const typeMapper: PluginTypeMapper = () => ({
+  [AUDIO_TYPE]: { component: AudioViewer },
+});